backend_api/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]

# For MongoDB integration
motor

# For Authentication
python-jose[cryptography]
passlib[bcrypt]

# For Ansible integration (to be added later)
# ansible-runner

